backend/requirements.txt
asgiref==3.4.1
async-timeout==4.0.3
autopep8==1.6.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==22.0.0
msgpack==1.0.8
packaging==24.1
psycopg2-binary==2.9.9
pycodestyle==2.8.0
PyJWT==2.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2021.3
redis==5.0.7
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.9.0
tzdata
pandas
requests==2.32.3
bs4==0.0.2
selenium==4.23.0
SQLAlchemy==2.0.31
pillow
xlrd
openpyxl
pymupdf
playwright
setuptools
